requirements.txt
aioredis==1.3.1
asgiref==3.2.7
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.6
django-crispy-forms==1.9.0
django-emoji==2.2.2
django-extensions==2.2.9
djangorestframework==3.11.0
Faker==4.1.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
model-bakery==1.1.0
more-itertools==8.2.0
msgpack==0.6.2
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==5.4.2
pytest-asyncio==0.12.0
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0
zope.interface==5.1.0
